requirements.txt
flask>=2.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
numpy>=1.21.0
pillow>=8.0.0
